🔥 chore(package.json): remove engines.node field as it is no longer needed
The engines.node field was removed from the package.json file as it is no longer needed. This field specifies the minimum version of Node.js required to run the application. Since the application no longer has any specific requirements on the version of Node.js, this field was removed to simplify the package.json file.