chore(changelog.yaml): remove changelog workflow file
The changelog.yaml file was removed as it is no longer needed. It was a workflow file that was responsible for releasing the package based on conventional commit messages. However, this workflow is no longer being used and has been deemed unnecessary. Therefore, the file was deleted to clean up the repository.