chore(changelog.yaml): add changelog generation workflow
This commit adds a new file `.github/workflows/changelog.yaml` which contains a workflow configuration for generating a changelog. The workflow is triggered on push and pull request events on the `develop` branch. It runs on the latest version of Ubuntu and consists of two steps.

The first step uses the `actions/checkout@v3` action to checkout the repository with a fetch depth of 0, ensuring that the entire commit history is available for generating the changelog.

The second step uses the `cocogitto/cocogitto-action@v3` action to perform a conventional commit check. This action checks the latest tag only and is configured to not release a new version. It also sets the git user and email to 'Ali Sait Teke' and 'alisaitteke@gmail.com' respectively.

No newline at end of file.