feat(pull-request.yml): add GitHub Actions workflow to automatically create a pull request on the develop branch
The pull-request.yml file was added to the .github/workflows directory. This file contains a GitHub Actions workflow that is triggered on pushes targeting the default branch (develop). It also allows the workflow to be manually triggered from the Actions tab.

The workflow consists of a single job named "release" that runs on the latest version of Ubuntu. The job includes a step that uses the GitHub CLI (gh) to create a pull request. The pull request is created with the base branch specified as "base_branch" and the branch to merge specified as "branch_to_merge". The pull request title is set to "Merge branch_to_merge into base_branch" and the body is set to "Created by GitHub action". The GITHUB_TOKEN secret is used for authentication.

No newline at end of file.